Move transactions to its folder
