+++ lib/index.ts
@@ -3,4 +3,3 @@ export { LibraWallet } from './wallet';
 export { Mnemonic } from './wallet/Mnemonic';
 export { Account, AccountState } from './wallet/Accounts';
 export * from './transaction';
-export * from './Transactions';
