+++ lib/transaction/index.ts
@@ -1,1 +1,2 @@
 export * from './Errors';
+export * from './Transactions';
